+++ dart/geodata/CHANGELOG.md
@@ -4,6 +4,7 @@ NOTE: Version 0.12.0 currently under development (0.12.0-dev.0).
 
 ⚠️ Breaking changes:
 * [Check conformance classes known by OGC API Features #169](https://github.com/navibyte/geospatial/issues/169)
+* Removed deprecated functions to create GeoJSON and OGC API Features clients.
 
 🛠 Maintenance:
 * Removed extra internal export files and made internal imports more excplicit.